daemon: introduce observer_finisher for delayed free
We introduce an observer finisher which can be registered in
the compartment struct. This finisher is called now in the method
compartment_notify_observers() after all observers have been executed
in the notify observer list. With this it is possible to register a
callback which can cleanly destroy the compartment object triggerd by
an observer.

Previously on container reload the reaload callback destroyed the
compartment object, but in the compartment_notify observers() method
the compartment object is accessed after each callback, even if it is
the last one. This caused the use-after-free error below.

=================================================================
==210==ERROR: AddressSanitizer: heap-use-after-free on address \
 0x5100000011b8 at pc 0x560d9ae6a3a5 bp 0x7fffd1ff8710 sp 0x7fffd1ff8700
READ of size 8 at 0x5100000011b8 thread T0
    #0 0x560d9ae6a3a4 in compartment_notify_observers daemon/compartment.c:1677
    #1 0x560d9ae6fd49 in compartment_sigchld_handle_helpers daemon/compartment.c:700
    #2 0x560d9ae70319 in compartment_sigchld_cb daemon/compartment.c:783
    #3 0x560d9af16329 in event_signal_handler common/event.c:780
    #4 0x560d9af16329 in event_loop common/event.c:851
    #5 0x560d9ae4e6ac in main daemon/main.c:146
    #6 0x7fe2bf99b863 in __libc_start_call_main ../sysdeps/nptl/libc_start_call_main.h:58
    #7 0x7fe2bf99b90a in __libc_start_main_impl ../csu/libc-start.c:389
    #8 0x560d9ae509d4 in _start (/usr/sbin/cmld+0x13e9d4)

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>
(cherry picked from commit f17d10b3a19dc2b6a05d5dd57c10b38df10e2707)
Signed-off-by: Felix Wruck <felix.wruck@aisec.fraunhofer.de>