daemon/c_net: Remove interface by given name
The removal of veth interface by the name stored in c_net_interface_t is
not reliable because the interface names might change during runtime.

Referencing the veth interface by ID also does not work because the
interface has to be removed in the c0 namespace, while the interface ID
is only valid in the rootns.

The function is changed to simply remove the interface with the name
given as argument for future use. The check if the interface is in use
before destroying it is also removed.

Signed-off-by: Maximilian Peltzer <mpeltzer@blackned.de>