daemon/control: Log files can be removed upon retrieval.
After a log file was send to the client succesfully
the log files is removed. Under consideration that
betold log file is not the latest and currently
used log file.

If log files shall be removed is controllable by the
caller of GET_LAST_LOG.

The whole process is locked to prevent interception of
other log file removal procedures.

Signed-off-by: David Sonntag <dsonntag@blackned.de>