common/cryptfs: Do not exit on key smaller than 64 byte
Because the migration from AES-128 to AES-256 requires the encryption
with AES-128 first, we can not exit when the key does not match the
AES-256 key size. Instead we print a warning that a smaller key has been
used. In the case of migration, the warning will appear first and later
the key sized used for re-encryption.

Signed-off-by: Maximilian Peltzer <mpeltzer@blackned.de>