daemon/compartment: Properly handle child exit error in
If the early child dies, respectively exits with error through sync
socket in compartment_start_post_clone_early_cb(), we do not need
to explicitly kill the early child anymore. Otherwise, the return
value/ exit error code would not be propagated in any case. Because
a race of SIGABORT could be trigger the exit child handler of the
early child.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>